# api/openapi/v1.json
{
  "components": {
    "schemas": {
...
      "ArtistRef": {
        "description": "A single performer on a track (a track can have several).",
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Id"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          }
        },
        "title": "ArtistRef",
        "type": "object"
      },
...
      "AuthError": {
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Track": {
        "properties": {
          "album": {
            "default": "",
            "title": "Album",
            "type": "string"
          },
          "album_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Album Id"
          },
          "artist": {
            "default": "",
            "title": "Artist",
            "type": "string"
          },
          "artist_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "integer"
              }
            ],
            "default": "",
            "title": "Artist Id"
          },
          "artists": {
            "items": {
              "$ref": "#/components/schemas/ArtistRef"
            },
            "title": "Artists",
            "type": "array"
          },
          "cover": {
            "default": "",
            "title": "Cover",
            "type": "string"
          },
          "duration_sec": {
            "default": 0,
            "title": "Duration Sec",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "preview_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preview Url"
          },
          "rank": {
            "default": 0,
            "title": "Rank",
            "type": "integer"
          },
          "release_date": {
            "default": "",
            "title": "Release Date",
            "type": "string"
          },
          "title": {
            "default": "",
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id"
        ],
        "title": "Track",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "sf_session": {
        "description": "HttpOnly JWT session cookie set by the login or registration endpoint.",
        "in": "cookie",
        "name": "sf_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "LoggeRythm API",
    "version": "1.0.0",
    "x-loggerythm-compatible-contract-versions": [
      "v1"
    ],
    "x-loggerythm-contract-version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/me/likes": {
      "get": {
        "operationId": "list_likes_api_me_likes_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Track"
                  },
                  "title": "Response List Likes Api Me Likes Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "List Likes",
        "tags": [
          "likes"
        ]
      }
    },
    "/api/me/likes/contains": {
      "get": {
        "operationId": "likes_contains_api_me_likes_contains_get",
        "parameters": [
          {
            "in": "query",
            "name": "ids",
            "required": false,
            "schema": {
              "default": "",
              "title": "Ids",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Likes Contains Api Me Likes Contains Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Likes Contains",
        "tags": [
          "likes"
        ]
      }
    },
    "/api/me/likes/{deezer_id}": {
      "delete": {
        "operationId": "remove_like_api_me_likes__deezer_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "deezer_id",
            "required": true,
            "schema": {
              "title": "Deezer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Remove Like",
        "tags": [
          "likes"
        ]
      },
      "put": {
        "operationId": "add_like_api_me_likes__deezer_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "deezer_id",
            "required": true,
            "schema": {
              "title": "Deezer Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Track"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The session cookie is missing, invalid, expired, or no longer identifies a user."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError"
                }
              }
            },
            "description": "The authenticated account is pending approval or lacks permission for this operation."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "sf_session": []
          }
        ],
        "summary": "Add Like",
        "tags": [
          "likes"
        ]
      }
    },
...
  }
}